Unwrap SoftCrashExceptions so we can actually deal with the failure
Calls to our remote process are wrapped in a "SoftCrashException"
in an attempt to prevent crashes; this exception derives from
OperationCanceledException. This backfires: our GraphQueryManager
does support reporting exceptions to an info bar, so this means anything
that goes wrong in the RPC then gets silently swallowed because we think
it's cancellation instead. Unwrapping makes things better again.

This commit should be reverted when we remove SoftCrashException in
https://github.com/dotnet/roslyn/issues/40476.