Deal with file mappings where the drive letter doesn't have a \
Path.Combine("C:", "Foo.txt") returns "C:Foo.txt" which reflects the
behavior that Foo.txt is relative not to C:\ but to the working
directory of C: which is C:Foo.txt. This isn't the behavior I expected
so this was broken. Add the extra directory slash in and do more testing
so we get this right.