Clean up some names in closure conversion (#41466)
We have two inaccurate terms left over from the closure
conversion rewrite: "lambda rewriter", which is now wrong
simply because it doesn't just rewrite lambdas, and "closure",
which was improperly conflated in the initial work to mean both
all nested functions, and the lowered form post-closure conversion
with all captures removed.

This change attempts to unify the naming for clarity. Local functions
and lambdas are now referred to as "nested functions" and "closure"
refers to the post-rewrite lowered method + environment(s).