Invoke generators in the CompilationTracker
This moves the invocation of generators into the CompilationTracker
so generators are properly included in all Compilations handed out
by the Workspace. This also allows us to take advantage of the
incremental update support of the compiler APIs.

The TrackedGeneratorDriver is a wrapper type to track some information
that right now is not tracked by GeneratorDriver for now; cleaning up
that API will be addressed in a future PR and the TrackedGeneratorDriver
type will go away.