Support ldftn
This change enables an address-of operator (&) to target a method group,
and for that result to be converted a function pointer type. Much like a
method-group-to-delegate conversion, an address-of-method-group
has no natural type, and must be target-typed to something with a
compatible function pointer type signature. Compatible is defined in
our function pointer spec here:
https://github.com/dotnet/csharplang/blob/master/proposals/function-pointers.md#function-pointer-conversions
An address of a method group, if valid, will load the function with the
ldftm CIL instruction, and this token can then be `calli`'d. This PR
does not implement support for `NativeCallableAttribute`, that will
come in a future PR. When that does, it will change the calling
convention that the method can be converted to.