Move VirtualCharsService and some internal SyntaxGenerator methods to shared layer
IVirtualCharsService is now a pure service (not ILanguageService) that is used in shared analyzer layer. A thin shim IVirtualCharsLanguageService has been added on top of IVirtualCharsService for clients that want MEF based discovery of this service. This follows the design used for syntax facts being split into pure service and a thin language service shim on top of it.