Fix race where we weren't writing LSIF output in the correct order
We had a race where we might have one thread trying to use vertices
that were still in the process of being written by another thread;
adding a lock to the shared LsifJsonWriter addresses this. To make the
use more clear InMemoryLsifJsonWriter is now renamed to
BatcingLsifJsonWriter and now accepts the underying object once so that
way there isn't the potential of misuse of different flushes going to
different writers which would once again break rules.

Unit tests are now added for the multi-document case which on my machine
was sufficient to reproduce the race.