Use init-only for records and adjust binding (#44582)
Changes records to generate init-only properties instead of get-only,
and modifies the `with` expression to treat arguments as assignments
to the left-hand side, with the safety rules of object initializers
(meaning init-only assignments are allowed).

Also changes records to generate a Clone method and a copy constructor.
The clone calls the copy constructor and the copy constructor does a
memberwise copy of the instance fields of the other type.

This change also tries to simplify and reuse more pieces of binding and
lowering from other initializer-based forms. The with-expression is now
bound as though it were an assignment to a field or member access on a
call to the Clone method.