Don't generate references for built-in operators
Since we didn't have a moniker for built-in operators, we were confusing
LSIF importing tools that weren't sure what to do here. They would
make up a moniker based on the first use, but that's not stable across
different runs.

We _could_ make up a moniker here, but it's not clear if finding
references for all uses of + in a project is really terribly useful.