Update the WorkCoordinatorTests
Reloading a project is now triggering work to happen because the
CompilationOptions won't be identical after the reload: we'll have a
new SyntaxTreeOptionsProvider which although would answer with the
same values, won't have object identity in the way the tests would
expect.

These APIs aren't used in Visual Studio so I don't think it's a big
deal.