Don't fetch output paths for some Intellisense projects
Web projects based on the legacy project system use special projects
called "intellisense projects" which are fake projects that are used
to act as a container for some contained languages. These are usually
created with the same IVsHierarchy as a regular project, which has an
unfortunate side effect. We get our build output paths from the
IVsHierarchy, so it would mean both the "real" project and the
Intellisense project would get the same output path.

In VS 2017, we would use that build output path to potentially convert
metadata references on disk back to project references. The ambiguous
path was problematic, but there was a bug: if the metadata reference
was converted to a project reference when there was only one project
(i.e. before you opened the file with a contained document), we
converted it but then left that project converted; if a later project
came along and outputted to the same output path the connection would
be left around. If the reference got added after the duplicate project,
it'd potentially stay a metadata reference.

In VS 2019, we rewrote the whole system that did
metadata-reference-to-project-reference conversion, and this time wrote
it "correctly" in that the end state didn't depend on the order of
events you took to get there; if you have two projects outputting the
same path, then we just don't convert a metadata reference to a project
reference.

The fix being taken here is to recognize when the projects we are being
handed are IVsIntellisenseProjects which means we can distinguish the
projects that don't need an output path from ones that do. This ensures
we don't end up with these projects having output paths at all, which
probably wouldn't end well since they don't correspond to the actual
code ultimately emitted to disk. We'll only ignore the project if
it also has a duplicate IVsHierarchy, to avoid breaking project types
(like database projects) that do create an IVsIntellisenseProject for
some files, but it's not a secondary project but rather the only
project.