Fix up nondeterminism in serializing naming style preferences
We have two members in TypeKind that point to the same value, Struct
and Structure. Because of this, Enum.ToString(), which under the covers
uses a binary search, isn't stable which one it will pick and it can
change if other TypeKinds are added. This ensures we keep using the
same string consistently.

We also have a few members in MethodKind that are also duplicates, but
it appears those have kept stable. To be safe, we now serialize
explicitly.

Fixes https://github.com/dotnet/roslyn/issues/44714