Fix nullable semantic model threading issue
Before taking the read lock on the node map, we check to ensure that no one else has done the work before us. After the lock is acquired, we would attempt to see if the Root was in the map and then exit if the work had already been done. However, this isn't precise enough for some cases, particularly field initializers. For FieldInitializers, the Root of the MemberSemanticModel is a VariableDeclaratorSyntax node. This node doesn't have a corresponding BoundNode in this case, however. Instead, the root BoundNode is a EqualsValueClauseSyntax node, so the existing check would return false, the root has not already been bound. We can replace this search with a simpler check of the count: since EnsureNullabilityAnalysisPerformedIfNecessary is called before attempting to do any individual node binding, we're guaranteed that if any nodes are in the map, they include the result of binding the root.