Support new function pointer metadata encoding
Adds support for the new extensible calling convention proposal in metadata. We bump the version of SRM (and therefore SCI) in order to be able to read the new calling convention bit, and then implement support for looking through the correct set of custom modifiers to be able to check calling convention equivalence. Order and duplication among the conventions does not matter for symbol equality. Modiifer order is copied when overriding as usual. For now, we have a flag to override checking for the RuntimeFeatures flag, as we don't yet have a runtime to test it on.