Completely obsolete the methods that take obsolete parameters
This obsoletes the methods to produce a SyntaxTree that either took
diagnostic options or the generated flag. We're electing to do this by
deleting the optional parameters to the API, and restoring the optional
parameters to the previous methods. This is a source-breaking change
if somebody was expicitly trying to use the old parameters with named
arguments, but since this is an intentional obsoletion that's probably
for the best, since anybody compiling against the new version will
know their code is broken.