Null annotate SymbolKey
The various reading methods when resolving keys can potentially return
null. The rule was as follows:

1. If the thing being written can never be null, I put a ! after the
   Read call.
2. If the thing being written could be null, no suppression was
   allowed.

Essentially this means that the reading code is not expected to be
robust in the face of nulls coming from incorrectly annotated code.

Some of the types (like SymbolKeyReader) are pooled. I elected to
leave the members annotated as they act when being used normally, and
suppressed null assignments when the object was in the pool.