Handle exception during teardown (#46727)
The `CancellationTokenSource.Cancel` method can throw and that needs to
be accounted for. This exception is not coming from our code but rather
it's coming from the framework code. It is attempting to access a
disposed handle when cancellation IO operations and that results in an
unhandled exception.

This can't really be prevented on our part. We just need to handle it.

closes #46724