Add a cache for AnalyzerConfigSets
The change to move .editorconfig off of syntax trees meant that we were
recomputing a bunch of information again and again, since previously
the cached syntax trees implicitly acted as a cache of this data.
This adds a quick cache back to avoid some extra overhead.

This cache is a bit more expensive than it probably needs to be, but
it's not too bad as the underlying AnalyerConfigSet has some logic
to avoid recreating duplicate arrays/dictionaries of the underlying
data, so much of the data is ultimately shared in the end.