Enable go to definition for source generated files
This allows go to definition (and similar features) to open the outputs
from a source generator as a "virtual" file. The file updates as
changes happen in near-real-time.

This commit is ignoring for now the problem of how we want to represent
source generated files as Documents in the Workspace; this isn't a big
deal bit a helper method (TryGetGeneratorAndHint) is added to easily
check if a SyntaxTree came from a generator, and if so gives you the
information you need to locate the generated file in a later run of
the generator. The majority of this pull request doesn't really change
once we have that working, so this isn't throw-away work. It does mean
though that we don't have a good way to link the generated file back
up to the generating project, so you only get a Miscellanous Files
experience in the generated document for now.

This is also doing a somewhat icky hack where we still have to put an
empty file onto disk, and fill it in once the file is open. The
challenge being that Visual Studio really doesn't give us any way
to have a moniker that isn't a file that passes through the system
at this point.