Switch from throwing NotImplementedException and return E_NOTIMPL
This is consistent with everything else in this class; I suspect this
was an oversight when the original implementer used implement interface
and missed updating this one. Oddly I saw this as a flagged exception
in an integration test run, so it's possible some code isn't handling
the CLR exception well.