Correctly get binder when a local function is the first thing in top level statements
We had a bug where nullable analysis was getting the incorrect binder when attempting to populate initial information about all top-level statements if the first thing in the top level was a local function. This ensures that we use the correct binder for these cases.

I've investigated other areas where we call `GetAdjustedNodePosition`, and I do not believe that any of them have this issue. It seems like this particular case is unique in that preferring the innermost binder was actually not what we wanted to do here, or the usage wasn't possible to hit with this scenario.