Mitigation for IOperation threading issue
This is a mitigation for the threading issue discussed in https://github.com/dotnet/roslyn/issues/35818. When building VS, there is a flaky threading issue where an interpolation with multiple parts can end up with different nodes on different threads, causing an optimization in `SetParentOperation(ImmutableArray)` to be invalid. This makes the optimization much narrower, ensuring that it is actually valid for this case.

The specific scenario here is an interplotated string with multiple parts. The steps are:
1. 2 threads attempt to get the children of an interpolated string operation at the same time.
2. Thread 1 gets the child for the first part of the string and caches it.
3. Thread 2 gets the child for the first part of the string and gets that same cached element.
4. Both threads attempt to get child 2. Neither hits the cache, so they get different nodes.
5. Both threads construct the array of interpolated string parts.
6. Thread 1 sets the parent of the interpolated string parts to be itself.
7. Thread 2 attempts to set the parent of the interpolated string parts to be itself. This returns early, because part 1 already has a parent, leaving part 2 with an unset parent.
8. Later, thread 2 asks for the parent of part 2. It is unset, violating the invariants, and we infinite loop.

As an additional mitigation step, we also ensure that the process of constructing an interpolated string's inserts does not hit the cache.  Since we never ask for an operation within a single expression, this will prevent 2 threads from seeing the same first element and differing subsequent elements.