Re-enable tests for use conditional expression for assignment
This requires changing a few baselines; it turns out that we had some
tests that (by name) said we should prefer using var and leaving a cast,
but the actual test was asserting the reverse: we didn't use var and
deleted the cast instead. This was happening deep in the simplification
pass, where we decided that we couldn't replace it with var because the
assignment to the local was causing a conversion. There wasn't actually
a conversion but instead the compiler was in an error case: the
refactoring had parenthesized the throw expression, which meant the
compiler gave up on type analysis entirely and was returning error
types. This was fixed in 9809f735c300e14f8351b062eb6df3f9f44331f5 and
so now the tests recognize we can replace with a var.