Separate build and test into separate jobs (#47249)
This makes several changes to our primary CI pipeline: 

1. Build and Test now occur on different machines. The test artifacts necessary to run tests are passed as pipeline artifacts between the jobs.
1. Every Job is now broken down into finer grained parts. For example Restore and Build will now run as separate tasks. This makes it easier to track certain analytical elements in our pipeline: what tasks are most flaky, when there are large time shifts in tasks, etc ... Today that is possible but requires digging through log files, now it will be evident in the UI in most cases. 
1. Removes four redundant builds from the job. 
1. Makes test retry faster because now it is just retrying tests, not lumping build in with it. This will be important as we embraced structured retries across the .NET infrastructure.

While this change does reduce our throughput slightly, that is not the goal here. The goal of this change is to get our workflow to be multi-machine, as well as getting us used to a multi-machine workflow, so that we can transition into Helix in the near future. 

This PR doesn't actually change much about how we build and test. The same commands of `-build, `-restore` and `-test` are being used to run the workflow. That means our workflow isn't optimal: we're still checking out sources and restoring on the test machines (wasted time). I will be changing that over the next week. Felt this was a significant milestone though and wanted to get it merged and then start the work of making this approach much more efficient.

This PR does change the emphasis of certain actions. For example the build jobs which feed into our test jobs are now about simply building the code as fast as possible. This means we don't run analyzers, bootstrap, etc ... Those actions are about validating the correctness of our build and are all covered in the correctness leg. The goal of building for unit tests is to do it as fast as possible so we can get the tests running.